Update bits saving

+++ scada.py
@@ -283,7 +283,7 @@ class Device:
                 result_code = struct.unpack('<H', data[4:6])[0]
 
                 for bit_idx in range(self.channels):
-                    bit_vals = ((result_code >> bit_idx) & 1).astype(int)
+                    bit_vals = (result_code >> bit_idx) & 1
                     samples.append(bit_vals)
 
                 errs = list(data[6:10])
